requirements.txt
django==2.1
mysqlclient==1.3.13
requests==2.19.1
ipaddress==1.0.22
djangorestframework==3.8.2
django-cors-headers==2.4.0
isort==4.3.4
